src/shim/pyproject.toml
[project]
name = "sem-video-shim"
version = "0.1.0"
description = "Video Capture Sidecar for JEOL SEM Linux Port"
authors = [{ name = "Antigravity", email = "antigravity@example.com" }]
dependencies = [
    "opencv-python",
    "pyqt6",
    "pyzmq",
    "numpy",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
